refactor(front): renames old files using kebab-cases
