--- client/src/components/Main/index.ts
@@ -1,3 +1,0 @@
-import { Main } from './Main';
-
-export default Main;
