--- client/src/components/Svgs/NavBarLogoSvg/NavBarLogoSvg.tsx
@@ -1,28 +1,0 @@
-import COLORS from 'styles/colors';
-
-interface NavBarLogoSvgProps {
-	className?: string;
-	size?: string;
-	color?: string;
-	backgroundColor?: string;
-}
-
-export const NavBarLogoSvg = ({
-	size = '38px',
-	color = COLORS.DARK.DEFAULT,
-	backgroundColor = COLORS.PRIMARY.DEFAULT
-}: NavBarLogoSvgProps) => {
-	return (
-		<div>
-			<svg width={size} height={size} viewBox="0 0 30 30" fill="none" xmlns="http://www.w3.org/2000/svg">
-				<rect width={30} height={30} rx="3" fill={backgroundColor}/>
-				<path
-					fill-rule="evenodd"
-					clip-rule="evenodd"
-					d="M9.32333 5.73099C8.0741 6.3909 6.40638 7.83699 5.61771 8.94509C4.82905 10.0532 3.69186 11.451 3.092 12.0509C1.25714 13.8857 1.7431 16.3737 4.63095 19.9319C7.78067 23.8121 10.858 25.5238 14.6855 25.5238C18.7329 25.5238 21.053 24.3154 24.75 20.2823C28.7633 15.9044 28.9632 14.1668 25.8333 10.8487C24.6417 9.58518 23.6667 8.34276 23.6667 8.08771C23.6667 7.83266 22.4583 6.91523 20.9812 6.0498C17.6173 4.07814 12.7064 3.9438 9.32333 5.73099ZM20.3114 7.60857C25.01 10.4958 25.6352 17.0342 21.599 21.071C19.6763 22.9931 18.9322 23.307 15.7497 23.5373C11.3903 23.8524 8.72348 22.7133 6.92762 19.7673C4.84205 16.3477 5.60286 10.5949 8.45357 8.22947C11.5736 5.63999 16.6727 5.37318 20.3114 7.60857ZM9.98448 8.98718C5.61338 11.9283 5.66538 18.3317 10.0736 19.8688C12.8278 20.8289 17.7028 20.7522 20.1981 19.7097C24.7085 17.8247 24.1452 10.9886 19.272 8.46842C16.6207 7.09723 12.4464 7.32999 9.98448 8.98718ZM13.1181 10.3274C14.7487 12.2923 11.6342 15.9546 9.69848 14.3481C8.47214 13.3304 8.58605 12.1282 10.0476 10.6667C11.521 9.19333 12.122 9.12709 13.1181 10.3274Z"
-					fill={color}
-				/>
-			</svg>
-		</div>
-	);
-};
