--- client/src/components/Svgs/NavBarLogoSvg/index.ts
@@ -1,3 +1,0 @@
-import { NavBarLogoSvg } from './NavBarLogoSvg';
-
-export default NavBarLogoSvg;
