+++ client/src/components/main/index.ts
@@ -1,0 +1,3 @@
+import { Main } from './main';
+
+export default Main;
