+++ client/src/components/main/main.tsx
@@ -7,8 +7,11 @@ interface MainProps {
 }
 
 export const Main = ({ children, className }: MainProps) => {
+	const marginLeftForNavBar = 'mt-16 ml-0 md:ml-14 md:mt-0';
 
-	const marginLeftForNavBar = 'mt-16 ml-0 md:ml-14 md:mt-0'
-
-	return <main className={twMerge(marginLeftForNavBar, 'flex flex-col h-full w-full', className)}>{children}</main>;
+	return (
+		<main className={twMerge(marginLeftForNavBar, 'flex h-full w-full flex-col', className)}>
+			{children}
+		</main>
+	);
 };
