+++ client/src/components/svgs/nav-bar-logo/index.ts
@@ -1,0 +1,3 @@
+import { NavBarLogoSvg } from './nav-bar-logo-svg';
+
+export default NavBarLogoSvg;
